Add stories folder inside the each component folder then split type definitions to utils folder
