--- src/components/ToggleSwtich/index.svelte
@@ -1,3 +1,0 @@
-<script lang="ts" context="module">
-    export * from './ToggleSwitch.svelte';
-</script>
