+++ src/components/ToggleSwtich/index.ts
@@ -1,0 +1,3 @@
+import ToggleSwitch from './ToggleSwitch.svelte';
+
+export default ToggleSwitch;
