+++ src/components/ToggleSwtich/stories/index.stories.svelte
@@ -1,6 +1,6 @@
 <script>
     import {Meta, Template, Story} from "@storybook/addon-svelte-csf";
-    import ToggleSwitch from "./ToggleSwitch.svelte";
+    import ToggleSwitch from "../ToggleSwitch.svelte";
   </script>
   
   <Meta
