+++ src/utils/types.svelte
@@ -1,0 +1,3 @@
+<script lang="ts" context="module">
+    export type Size = 'large' | 'medium' | 'small';
+</script>
